add service impl

+++ seckill/src/main/java/club/lightingsummer/seckill/seckill/dao/RecordMapper.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Repository;
 public interface RecordMapper {
 
     // 插入秒杀记录
-    int insertSuccessKilled(@Param("seckillId") Integer seckillId, @Param("seckillPhone") Long seckillPhone);
+    int insertSuccessKilled(@Param("seckillId") long seckillId, @Param("seckillPhone") long seckillPhone);
 
     // 根据秒杀id查询
-    Record selectById(@Param("seckillId") Integer seckillId);
+    Record selectById(@Param("seckillId") long seckillId);
 }
